+++ server/db/index.ts
@@ -16,6 +16,7 @@ import * as staffSchema from './schema/staff'
 import * as chatSchema from './schema/chat'
 import * as announcementSchema from './schema/announcement'
 import * as pollSchema from './schema/poll'
+import * as pushPreferencesSchema from './schema/push-preferences'
 
 const connectionString = process.env.DATABASE_URL!
 
@@ -39,5 +40,6 @@ export const db = drizzle(client, {
     ...chatSchema,
     ...announcementSchema,
     ...pollSchema,
+    ...pushPreferencesSchema,
   },
 })